Feat: Add BinaryDataAccessor
The added binary data accessor is a copy of the FileDataAccessor from the solid-community-server source, which will be
taken as a basis to implement the BinaryDataAccessor. The plan is to pass the filesystem object as a constructor
argument to the accessor, such that other binary backends only have to implement the basic filesystem interface. This
separation decouples the binary "backend" (Fs, IPFS,...) from the data accessor's solid specific implementation details.
And should therefore help to keep in sync with upcoming changes to the accessor.